Organize event handlers into separate folders
