--- macros/blueprint-proc-macro/src/blueprint.rs
@@ -1,1 +1,0 @@
-
