--- macros/blueprint-proc-macro/src/contracts/mod.rs
@@ -1,1 +1,0 @@
-pub mod solidity;
