--- macros/blueprint-proc-macro/src/contracts/solidity.rs
@@ -1,34 +1,0 @@
-#[proc_macro]
-pub fn generate_solidity(input: TokenStream) -> TokenStream {
-    let input = parse_macro_input!(input as syn::ItemMod);
-    let mod_name = &input.ident;
-
-    // Analyze the module content to find job and report definitions
-    // Generate Solidity code based on these definitions
-
-    let solidity_code = generate_solidity_contract(&input);
-
-    let expanded = quote! {
-        const SOLIDITY_CONTRACT: &str = #solidity_code;
-
-        fn get_solidity_contract() -> &'static str {
-            SOLIDITY_CONTRACT
-        }
-    };
-
-    TokenStream::from(expanded)
-}
-
-fn generate_solidity_contract(module: &syn::ItemMod) -> String {
-    let mut contract = String::new();
-
-    contract.push_str("pragma solidity ^0.8.20;\n\n");
-    contract.push_str(&format!("contract {} {{\n", module.ident));
-
-    // Generate contract content based on job and report definitions
-    // This would involve traversing the AST of the Rust module
-
-    contract.push_str("}\n");
-
-    contract
-}
