+++ macros/blueprint-proc-macro/src/shared.rs
@@ -1,0 +1,151 @@
+use gadget_blueprint_proc_macro_core::FieldType;
+use syn::{Ident, Type};
+
+/// Convert a `snake_case` string to `PascalCase`
+pub fn pascal_case(s: &str) -> String {
+    s.split('_')
+        .map(|word| {
+            let mut c = word.chars();
+            match c.next() {
+                None => String::new(),
+                Some(f) => f.to_uppercase().collect::<String>() + c.as_str(),
+            }
+        })
+        .collect()
+}
+
+pub fn ident_to_field_type(ident: &Ident) -> syn::Result<FieldType> {
+    match ident.to_string().as_str() {
+        "u8" => Ok(FieldType::Uint8),
+        "u16" => Ok(FieldType::Uint16),
+        "u32" => Ok(FieldType::Uint32),
+        "u64" => Ok(FieldType::Uint64),
+        "i8" => Ok(FieldType::Int8),
+        "i16" => Ok(FieldType::Int16),
+        "i32" => Ok(FieldType::Int32),
+        "i64" => Ok(FieldType::Int64),
+        "u128" => Ok(FieldType::Uint128),
+        "i128" => Ok(FieldType::Int128),
+        "f64" => Ok(FieldType::Float64),
+        "bool" => Ok(FieldType::Bool),
+        "String" => Ok(FieldType::String),
+        "Bytes" => Ok(FieldType::Bytes),
+        "AccountId" => Ok(FieldType::AccountId),
+        _ => Err(syn::Error::new_spanned(ident, "unsupported type")),
+    }
+}
+
+pub fn type_to_field_type(ty: &Type) -> syn::Result<FieldType> {
+    match ty {
+        Type::Array(_) => Err(syn::Error::new_spanned(ty, "TODO: support arrays")),
+        Type::Path(inner) => path_to_field_type(&inner.path),
+        Type::Reference(type_reference) => type_to_field_type(&type_reference.elem),
+        _ => Err(syn::Error::new_spanned(ty, "unsupported type")),
+    }
+}
+
+pub fn path_to_field_type(path: &syn::Path) -> syn::Result<FieldType> {
+    // take the last segment of the path
+    let seg = &path
+        .segments
+        .last()
+        .ok_or_else(|| syn::Error::new_spanned(path, "path must have at least one segment"))?;
+    let ident = &seg.ident;
+    let args = &seg.arguments;
+    match args {
+        syn::PathArguments::None => {
+            match ident_to_field_type(ident) {
+                Ok(field_type) => Ok(field_type),
+                Err(_) => {
+                    // Assume it's a custom struct if it's not a known type
+                    Ok(FieldType::Struct(ident.to_string(), Vec::new()))
+                }
+            }
+        }
+        // Support for Vec<T> where T is a simple type
+        syn::PathArguments::AngleBracketed(inner) if ident.eq("Vec") && inner.args.len() == 1 => {
+            let inner_arg = &inner.args[0];
+            if let syn::GenericArgument::Type(inner_ty) = inner_arg {
+                let inner_type = type_to_field_type(inner_ty)?;
+                match inner_type {
+                    FieldType::Uint8 => Ok(FieldType::Bytes),
+                    others => Ok(FieldType::List(Box::new(others))),
+                }
+            } else {
+                Err(syn::Error::new_spanned(
+                    inner_arg,
+                    "unsupported complex type",
+                ))
+            }
+        }
+        // Support for Option<T> where T is a simple type
+        syn::PathArguments::AngleBracketed(inner)
+            if ident.eq("Option") && inner.args.len() == 1 =>
+        {
+            let inner_arg = &inner.args[0];
+            if let syn::GenericArgument::Type(inner_ty) = inner_arg {
+                let inner_type = type_to_field_type(inner_ty)?;
+                Ok(FieldType::Optional(Box::new(inner_type)))
+            } else {
+                Err(syn::Error::new_spanned(
+                    inner_arg,
+                    "unsupported complex type",
+                ))
+            }
+        }
+        // Support for Result<T, E> where T is a simple type
+        syn::PathArguments::AngleBracketed(inner) if ident.eq("Result") => {
+            let inner_arg = &inner.args[0];
+            if let syn::GenericArgument::Type(inner_ty) = inner_arg {
+                let inner_type = type_to_field_type(inner_ty)?;
+                Ok(inner_type)
+            } else {
+                Err(syn::Error::new_spanned(
+                    inner_arg,
+                    "unsupported complex type",
+                ))
+            }
+        }
+        syn::PathArguments::Parenthesized(_) => Err(syn::Error::new_spanned(
+            args,
+            "unsupported parenthesized arguments",
+        )),
+        _ => Err(syn::Error::new_spanned(args, "unsupported complex type")),
+    }
+}
+
+#[allow(dead_code)]
+pub fn parse_struct_fields(fields: &syn::Fields) -> syn::Result<Vec<(String, FieldType)>> {
+    fields
+        .iter()
+        .map(|field| {
+            let name = field
+                .ident
+                .as_ref()
+                .ok_or_else(|| syn::Error::new_spanned(field, "Unnamed fields are not supported"))?
+                .to_string();
+            let field_type = type_to_field_type(&field.ty)?;
+            Ok((name, field_type))
+        })
+        .collect()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn pascal_case_works() {
+        let input = [
+            "hello_world",
+            "keygen",
+            "_internal_function",
+            "cggmp21_sign",
+        ];
+        let expected = ["HelloWorld", "Keygen", "InternalFunction", "Cggmp21Sign"];
+
+        for (i, e) in input.iter().zip(expected.iter()) {
+            assert_eq!(pascal_case(i), *e);
+        }
+    }
+}
